Use viewModel in TweetList for open Twitter function
