+++ brain-marks/Tweets/TweetListViewModel.swift
@@ -18,4 +18,18 @@ final class TweetListViewModel: ObservableObject {
             }
         }
     }
+    
+    /// Opens Twitter app when tapping on tweet card
+    func openTwitter(tweetID:String, authorUsername:String) {
+        let appURL = NSURL(string: "twitter://\(authorUsername)/status/\(tweetID)?s=21")!
+        let webURL = NSURL(string: "https://twitter.com/\(authorUsername)/status/\(tweetID)?s=21")!
+        
+        let application = UIApplication.shared
+        
+        if application.canOpenURL(appURL as URL) {
+            application.open(appURL as URL)
+        } else {
+            application.open(webURL as URL)
+        }
+    }
 }
